Has variable "brute forcing" and nice prints
Variable brute forcing is if the contours are not well defined enough due to image bluring it will force an extra rectangle with the same size of the third onto the fours digit, and if it cannot determine it still it throws an error saying provide a better image. Also now only shows final images and the temperature.